HuggingFaceDatasets

Documentation for HuggingFaceDatasets.

HuggingFaceDatasets.Dataset

HuggingFaceDatasets.DatasetDict

HuggingFaceDatasets.py2jl

HuggingFaceDatasets.Dataset — Type

Dataset(pydataset; transform = py2jl)

A Julia wrapper around the objects of the python datasets.Dataset class.

The transform is applied after datasets' one. The py2jl default converts python types to julia types.

Provides:

1-based indexing.

set_transform! julia method.

All python class' methods from datasets.Dataset.

See also load_dataset and DatasetDict.

HuggingFaceDatasets.DatasetDict — Type

DatasetDict(pydatasetdict::Py; transform = py2jl)

A DatasetDict is a dictionary of Datasets. It is a wrapper around a datasets.DatasetDict object.

The transform is applied to each Dataset. The py2jl default converts python types to julia types.

See also load_dataset and Dataset.

HuggingFaceDatasets.py2jl — Function

py2jl(x)

Convert Python types to Julia types applying pyconvert recursively.