API

Index

HuggingFaceDatasets.Dataset

HuggingFaceDatasets.DatasetDict

HuggingFaceDatasets.load_dataset

HuggingFaceDatasets.numpy2jl

HuggingFaceDatasets.py2jl

HuggingFaceDatasets.set_format!

HuggingFaceDatasets.set_format!

HuggingFaceDatasets.set_jltransform!

HuggingFaceDatasets.set_jltransform!

HuggingFaceDatasets.with_format

HuggingFaceDatasets.with_format

HuggingFaceDatasets.with_jltransform

HuggingFaceDatasets.with_jltransform

Docs

HuggingFaceDatasets.Dataset — Type

Dataset

A Julia wrapper around an object of the python datasets.Dataset class.

Provides:

1-based indexing.

All python class' methods from datasets.Dataset.

See also load_dataset and DatasetDict.

HuggingFaceDatasets.DatasetDict — Type

DatasetDict(pydatasetdict::Py; transform = identity)

A DatasetDict is a dictionary of Datasets. It is a wrapper around a datasets.DatasetDict object.

The transform is applied to each Dataset. The py2jl transform provided by this package converts python types to julia types.

See also load_dataset and Dataset.

HuggingFaceDatasets.load_dataset — Method

load_dataset(args...; kws...)

Load a dataset from the HuggingFace Datasets library.

All arguments are passed to the python function datasets.load_dataset. See the documentation here.

Returns a DatasetDict or a Dataset depending on the split argument.

Use the dataset.with_format("julia") method to lazily convert the observation from the dataset to julia types.

Examples

Without a split argument, a DatasetDict is returned:

julia> d = load_dataset("glue", "sst2")
DatasetDict({
    train: Dataset({
        features: ['sentence', 'label', 'idx'],
        num_rows: 67349
    })
    validation: Dataset({
        features: ['sentence', 'label', 'idx'],
        num_rows: 872
    })
    test: Dataset({
        features: ['sentence', 'label', 'idx'],
        num_rows: 1821
    })
})

julia> d["train"]
Dataset({
    features: ['sentence', 'label', 'idx'],
    num_rows: 67349
})

Selecting a split returns a Dataset instead. We also apply the "julia" format.

julia> mnist = load_dataset("mnist", split="train").with_format("julia")
Dataset({
    features: ['image', 'label'],
    num_rows: 60000
})

julia> mnist[1]
Dict{String, Any} with 2 entries:
  "label" => 5
  "image" => Gray{N0f8}[Gray{N0f8}(0.0) Gray{N0f8}(0.0) … Gray{N0f8}(0.0) Gray{N0f8}(0.0); Gray{N0f8}(0.0) Gray{N0f8}(0.0) … Gray{N0f…

HuggingFaceDatasets.numpy2jl — Method

numpy2jl(x)

Convert a numpy array to a Julia array using DLPack. The conversion is copyless, and mutations to the Julia array are reflected in the numpy array.

HuggingFaceDatasets.py2jl — Function

py2jl(x)

Convert Python types to Julia types applying pyconvert recursively.

HuggingFaceDatasets.set_format! — Method

set_format!(ds::Dataset, format)

Set the format of ds to format. Mutating version of with_format.

HuggingFaceDatasets.set_format! — Method

set_format!(d::DatasetDict, format)

Set the format of d to format. Mutating version of with_format.

HuggingFaceDatasets.set_jltransform! — Method

set_jltransform!(ds::Dataset, transform)
set_jltransform!(transform, ds::Dataset)

Set the julia transform of ds to transform. Mutating version of with_jltransform.

HuggingFaceDatasets.set_jltransform! — Method

set_jltransform!(d::DatasetDict, transform)
set_jltransform!(transform, d::DatasetDict)

Set the transform of d to transform. Mutating version of with_jltransform.

HuggingFaceDatasets.with_format — Method

with_format(ds::Dataset, format)

Return a copy of ds with the format set to format. If format is "julia", the returned dataset will be transformed with py2jl and copyless conversion from python types will be used when possible.

See also set_format!.

Examples

julia> ds = load_dataset("mnist", split="test");

julia> ds[1]
Python dict: {'image': <PIL.PngImagePlugin.PngImageFile image mode=L size=28x28 at 0x2B5B4C1F0>, 'label': 7}

julia> ds = with_format(ds, "julia");

julia> ds[1]
Dict{String, Any} with 2 entries:
  "label" => 7
  "image" => UInt8[0x00 0x00 … 0x00 0x00; 0x00 0x00 … 0x00 0x00; … ; 0x00 0x00 … 0x00 0x00; 0x00 0x00 … 0x00 0x00]

HuggingFaceDatasets.with_format — Method

with_format(d::DatasetDict, format)

Return a copy of d with the format set to format. If format is "julia", the returned dataset will be transformed with py2jl and copyless conversion from python types will be used when possible.

HuggingFaceDatasets.with_jltransform — Method

with_jltransform(ds::Dataset, transform)
with_jltransform(transform, ds::Dataset)

Return a copy of ds with the julia transform set to transform. The transform applies when indexing, e.g. ds[1] or ds[1:2].

The transform is always applied to a batch of data, even if the index is a single integer. That is, ds[1] is equivalent to ds[1:1] from the point of view of the transform.

The julia transform is applied after the python transform (if any). The python transform can be set with ds.set_transform(pytransform).

If transform is nothing or identity, the returned dataset will not be transformed.

See also set_jltransform! for the mutating version.

HuggingFaceDatasets.with_jltransform — Method

" withjltransform(d::DatasetDict, transform) withjltransform(transform, d::DatasetDict)

Return a copy of d with the julia transform applied to each Dataset.